+++ x/erc20/genesis.go
@@ -37,11 +37,6 @@ func InitGenesis(
 		}
 	}
 
-	for _, precompile := range data.NativePrecompiles {
-		if err := k.EnableNativePrecompile(ctx, common.HexToAddress(precompile)); err != nil {
-			panic(fmt.Errorf("error registering native precompiles %s", err))
-		}
-	}
 	for _, precompile := range data.DynamicPrecompiles {
 		if err := k.EnableDynamicPrecompile(ctx, common.HexToAddress(precompile)); err != nil {
 			panic(fmt.Errorf("error registering dynamic precompiles %s", err))
@@ -66,7 +61,6 @@ func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 		Params:             k.GetParams(ctx),
 		TokenPairs:         k.GetTokenPairs(ctx),
 		Allowances:         k.GetAllowances(ctx),
-		NativePrecompiles:  k.GetNativePrecompiles(ctx),
 		DynamicPrecompiles: k.GetDynamicPrecompiles(ctx),
 	}
 }